+++ packages/blockstore-core/src/memory.ts
@@ -1,14 +1,19 @@
 import { NotFoundError } from 'interface-store'
+import all from 'it-all'
 import { base32 } from 'multiformats/bases/base32'
 import { CID } from 'multiformats/cid'
 import * as raw from 'multiformats/codecs/raw'
 import * as Digest from 'multiformats/hashes/digest'
 import { BaseBlockstore } from './base.js'
 import type { Pair } from 'interface-blockstore'
-import type { AbortOptions, Await, AwaitIterable } from 'interface-store'
+import type { AbortOptions, Await, AwaitGenerator, AwaitIterable } from 'interface-store'
+
+function isPromise <T> (p?: any): p is Promise<T> {
+  return typeof p?.then === 'function'
+}
 
 export class MemoryBlockstore extends BaseBlockstore {
-  private readonly data: Map<string, Uint8Array>
+  private readonly data: Map<string, Uint8Array[]>
 
   constructor () {
     super()
@@ -16,14 +21,37 @@ export class MemoryBlockstore extends BaseBlockstore {
     this.data = new Map()
   }
 
-  put (key: CID, val: Uint8Array, options?: AbortOptions): Await<CID> {
+  put (key: CID, val: Uint8Array | AwaitIterable<Uint8Array>, options?: AbortOptions): Await<CID> {
     options?.signal?.throwIfAborted()
+
+    let buf: Uint8Array[]
+
+    if (val instanceof Uint8Array) {
+      buf = [val]
+    } else {
+      const result = all(val)
+
+      if (isPromise<Uint8Array[]>(result)) {
+        return result.then(val => {
+          return this._put(key, val, options)
+        })
+      } else {
+        buf = result
+      }
+    }
+
+    return this._put(key, buf, options)
+  }
+
+  private _put (key: CID, val: Uint8Array[], options?: AbortOptions): Await<CID> {
+    options?.signal?.throwIfAborted()
+
     this.data.set(base32.encode(key.multihash.bytes), val)
 
     return key
   }
 
-  get (key: CID, options?: AbortOptions): Await<Uint8Array> {
+  * get (key: CID, options?: AbortOptions): AwaitGenerator<Uint8Array> {
     options?.signal?.throwIfAborted()
     const buf = this.data.get(base32.encode(key.multihash.bytes))
 
@@ -31,7 +59,7 @@ export class MemoryBlockstore extends BaseBlockstore {
       throw new NotFoundError()
     }
 
-    return buf
+    yield * buf
   }
 
   has (key: CID, options?: AbortOptions): Await<boolean> {
@@ -44,13 +72,15 @@ export class MemoryBlockstore extends BaseBlockstore {
     this.data.delete(base32.encode(key.multihash.bytes))
   }
 
-  async * getAll (options?: AbortOptions): AwaitIterable<Pair> {
+  * getAll (options?: AbortOptions): AwaitGenerator<Pair> {
     options?.signal?.throwIfAborted()
 
     for (const [key, value] of this.data.entries()) {
       yield {
         cid: CID.createV1(raw.code, Digest.decode(base32.decode(key))),
-        block: value
+        bytes: (async function * () {
+          yield * value
+        })()
       }
       options?.signal?.throwIfAborted()
     }