+++ packages/blockstore-core/CHANGELOG.md
@@ -1,3 +1,47 @@
+## [blockstore-core-v6.1.2](https://github.com/ipfs/js-stores/compare/blockstore-core-6.1.1...blockstore-core-6.1.2) (2026-01-08)
+
+### Bug Fixes
+
+* identity supports async get and getAll ([#365](https://github.com/ipfs/js-stores/issues/365)) ([a1b95fc](https://github.com/ipfs/js-stores/commit/a1b95fc173c066d69bf0f1145d7b458941065be6))
+
+## [blockstore-core-v6.1.1](https://github.com/ipfs/js-stores/compare/blockstore-core-6.1.0...blockstore-core-6.1.1) (2025-10-09)
+
+### Bug Fixes
+
+* update error name ([304ec71](https://github.com/ipfs/js-stores/commit/304ec7142b16ad0cd70537d48eb69294faacffe5))
+
+## [blockstore-core-v6.1.0](https://github.com/ipfs/js-stores/compare/blockstore-core-6.0.2...blockstore-core-6.1.0) (2025-10-09)
+
+### Features
+
+* enforce max digest length ([#360](https://github.com/ipfs/js-stores/issues/360)) ([c5b039a](https://github.com/ipfs/js-stores/commit/c5b039ae248cce82bf02fbf0383da8e65e140786))
+
+## [blockstore-core-v6.0.2](https://github.com/ipfs/js-stores/compare/blockstore-core-6.0.1...blockstore-core-6.0.2) (2025-10-03)
+
+### Dependencies
+
+* bump level from 8.0.1 to 10.0.0 ([#356](https://github.com/ipfs/js-stores/issues/356)) ([c0ec61f](https://github.com/ipfs/js-stores/commit/c0ec61fe965e3bad9d607a0bd3a3c750f00f41d0))
+
+## [blockstore-core-v6.0.1](https://github.com/ipfs/js-stores/compare/blockstore-core-6.0.0...blockstore-core-6.0.1) (2025-10-03)
+
+### Bug Fixes
+
+* update sibling deps ([3f73b3d](https://github.com/ipfs/js-stores/commit/3f73b3d53ea2d86d0f5c3f06785c0bfc30e8b5e9))
+
+## [blockstore-core-v6.0.0](https://github.com/ipfs/js-stores/compare/blockstore-core-5.0.4...blockstore-core-6.0.0) (2025-10-03)
+
+### ⚠ BREAKING CHANGES
+
+* blockstore.get and similar now return streams of bytes
+
+### Features
+
+* streaming blockstores ([#358](https://github.com/ipfs/js-stores/issues/358)) ([4dbb136](https://github.com/ipfs/js-stores/commit/4dbb1362d20fc87fcdd261568dca297972f9bc08))
+
+### Dependencies
+
+* bump @libp2p/logger ([#359](https://github.com/ipfs/js-stores/issues/359)) ([edb5a1f](https://github.com/ipfs/js-stores/commit/edb5a1f8b575a27ad28bc2e1c4e4d52e1f114ebc))
+
 ## [blockstore-core-v5.0.4](https://github.com/ipfs/js-stores/compare/blockstore-core-5.0.3...blockstore-core-5.0.4) (2025-05-28)
 
 ### Bug Fixes