+++ packages/blockstore-s3/CHANGELOG.md
@@ -1,3 +1,29 @@
+## [blockstore-s3-v3.0.1](https://github.com/ipfs/js-stores/compare/blockstore-s3-3.0.0...blockstore-s3-3.0.1) (2025-10-03)
+
+### Bug Fixes
+
+* update sibling deps ([3f73b3d](https://github.com/ipfs/js-stores/commit/3f73b3d53ea2d86d0f5c3f06785c0bfc30e8b5e9))
+
+## [blockstore-s3-v3.0.0](https://github.com/ipfs/js-stores/compare/blockstore-s3-2.0.5...blockstore-s3-3.0.0) (2025-10-03)
+
+### ⚠ BREAKING CHANGES
+
+* blockstore.get and similar now return streams of bytes
+
+### Features
+
+* streaming blockstores ([#358](https://github.com/ipfs/js-stores/issues/358)) ([4dbb136](https://github.com/ipfs/js-stores/commit/4dbb1362d20fc87fcdd261568dca297972f9bc08))
+
+### Trivial Changes
+
+* bump sinon from 20.0.0 to 21.0.0 ([#351](https://github.com/ipfs/js-stores/issues/351)) ([f24dd4f](https://github.com/ipfs/js-stores/commit/f24dd4f9b6eb681b1f9652409a558a8bcc50f4a4))
+
+## [blockstore-s3-v2.0.5](https://github.com/ipfs/js-stores/compare/blockstore-s3-2.0.4...blockstore-s3-2.0.5) (2025-09-02)
+
+### Bug Fixes
+
+* deprecate blockstore-level and datastore-fs ([#353](https://github.com/ipfs/js-stores/issues/353)) ([ebc7912](https://github.com/ipfs/js-stores/commit/ebc7912696d5bd9dc991ece5f0c0d4acfb1f9400))
+
 ## [blockstore-s3-v2.0.4](https://github.com/ipfs/js-stores/compare/blockstore-s3-2.0.3...blockstore-s3-2.0.4) (2025-05-28)
 
 ### Bug Fixes