+++ packages/interface-datastore-tests/CHANGELOG.md
@@ -1,3 +1,15 @@
+## [interface-datastore-tests-v6.0.5](https://github.com/ipfs/js-stores/compare/interface-datastore-tests-6.0.4...interface-datastore-tests-6.0.5) (2025-10-03)
+
+### Dependencies
+
+* bump level from 8.0.1 to 10.0.0 ([#356](https://github.com/ipfs/js-stores/issues/356)) ([c0ec61f](https://github.com/ipfs/js-stores/commit/c0ec61fe965e3bad9d607a0bd3a3c750f00f41d0))
+
+## [interface-datastore-tests-v6.0.4](https://github.com/ipfs/js-stores/compare/interface-datastore-tests-6.0.3...interface-datastore-tests-6.0.4) (2025-10-03)
+
+### Bug Fixes
+
+* update sibling deps ([3f73b3d](https://github.com/ipfs/js-stores/commit/3f73b3d53ea2d86d0f5c3f06785c0bfc30e8b5e9))
+
 ## [interface-datastore-tests-v6.0.3](https://github.com/ipfs/js-stores/compare/interface-datastore-tests-6.0.2...interface-datastore-tests-6.0.3) (2025-05-28)
 
 ### Bug Fixes